% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{chunk_size}
\alias{chunk_size}
\title{Ajuste del tamaño del texto en chunks}
\usage{
chunk_size(...)
}
\arguments{
\item{...}{path al documento .Rmd}
}
\description{
Ajuste del tamaño del texto en chunks
}
