% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-codigos_ensenanza.R
\docType{data}
\name{codigos_ensenanza}
\alias{codigos_ensenanza}
\title{Codigos de enseñanza}
\format{data.frame con 44 filas (codigos de enseñanza) y 6 variables

\describe{
  \item{cod_ense}{Codigo de enseñanza}
  \item{descripcion}{Descripción de codigo de enseñanza}
  \item{cod_ense2}{Codigo de enseñanza agrupados 2, distingue educación especial}
  \item{cod_ense2_lab}{Descripción de codigo de enseñanza 2}
  \item{cod_ense3}{Codigo de enseñanza agrupados 3}
  \item{cod_ense3_lab}{Descripción de codigo de enseñanza 3}
}}
\usage{
codigos_ensenanza
}
\description{
Descripción y equivalencias entre codigos de enseñanza utilizados por
MINEDUC de Chile para la clasificación de escuelas.
}
\keyword{datasets}
