% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{collapse_chr}
\alias{collapse_chr}
\title{Colapso de strings

Variable útil para colapsar en una sola casilla _list variables_ en un
tibble que quiera ser mostrad en una tabla}
\usage{
collapse_chr(.data, .var, collapse = ", ")
}
\arguments{
\item{.data}{tibble}

\item{.var}{variable name}

\item{collapse}{string, string con el que se unirán los textos. Por defecto
", ".}
}
\description{
Colapso de strings

Variable útil para colapsar en una sola casilla _list variables_ en un
tibble que quiera ser mostrad en una tabla
}
