% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edad_rut.R
\name{edad_rut}
\alias{edad_rut}
\title{Estimación de edad a partir de RUT}
\usage{
edad_rut(.rut)
}
\arguments{
\item{.rut}{integer. RUT sin dígito verificador.}
}
\value{
integer. Edad estimada de la persona.
}
\description{
Estimación de edad a partir de RUT
}
\examples{
edad_rut(13456789)

}
