% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_cod_depe2}
\alias{edu_fct_cod_depe2}
\title{Factor de código de dependencia del establecimiento agrupado}
\usage{
edu_fct_cod_depe2(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{COD_DEPE2}.}
}
\value{
A factor with 5 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Código de Dependencia del Establecimiento (agrupado).
}
\seealso{
\url{https://centroestudios.mineduc.cl/datos-abiertos/}
}
