% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_ens}
\alias{edu_fct_ens}
\title{Factor de código de enseñanza en 13 niveles}
\usage{
edu_fct_ens(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{COD_DEPE2}.}
}
\value{
A factor with 13 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Recodificación del Código de Enseñanza en 13 niveles.
}
\seealso{
\url{https://centroestudios.mineduc.cl/datos-abiertos/}
}
