% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_rural_rbd}
\alias{edu_fct_rural_rbd}
\title{Factor de ruralidad}
\usage{
edu_fct_rural_rbd(x)
}
\arguments{
\item{x}{Un vector de c\u00f3digos, t\u00edpicamente variable `RURAL_RBD`.}
}
\value{
A factor with 2 levels, with an attached `"label"` attribute.
}
\description{
Recodificaci\u00f3n para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
\u00cdndice de ruralidad del establecimiento.
}
\seealso{
[https://centroestudios.mineduc.cl/datos-abiertos/]
}
