% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{fct_reorder_cat}
\alias{fct_reorder_cat}
\title{Orden de preguntas según valor de categoría}
\usage{
fct_reorder_cat(.f, .cat, .val, cat_orden, .desc = FALSE)
}
\arguments{
\item{.f}{factor. Factor que quiere ser ordenado.}

\item{.cat}{chr o factor. Variable con categorías de respuesta.}

\item{.val}{numeric. Valor de respuesta para cada categoría de respuesta}

\item{cat_orden}{string. Nombre de la categoría de respuesta de interés.}

\item{.desc}{logical. Controla si el orden es ascendente o descendente. Por defecto
.desc = FALSE.}
}
\value{
factor
}
\description{
Wrapper al rededor de `forcats::fct_reorder2` para ordenar factor `.f` según
según una catogoría de respuesta de `.x` de interès `cat_orden` a partir del valor en `.y`.
}
