% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_print.R
\name{format_clp}
\alias{format_clp}
\title{Dinero en formato CLP}
\usage{
format_clp(x, digits = 0)
}
\arguments{
\item{x}{numeric}

\item{digits}{cantidad de dígitos del número entero que se quiere obtener.
Por defecto digits = 0.}
}
\value{
character
}
\description{
Transforma número a texto como \code{CLP} con separación
de miles con \code{.} y decimales con \verb{,}.
}
\examples{

format_clp(1000000)

}
