% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_print.R
\name{format_dinero}
\alias{format_dinero}
\title{Dinero en formato $}
\usage{
format_dinero(x, digits = 0)

format_dinero(x, digits = 0)
}
\arguments{
\item{x}{numeric}

\item{digits}{cantidad de dígitos del número entero que se quiere obtener.
Por defecto digits = 0.}
}
\value{
character

character
}
\description{
Transforma número a texto como `$` con separación
de miles con `.` y decimales con `,`.

Transforma número a texto con separación
de miles con `.` y decimales con `,`.
}
\examples{

format_dinero(1000000)


format_dinero(1000000)

}
