% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_print.R
\name{format_num}
\alias{format_num}
\title{Número en formato para texto}
\usage{
format_num(x, digits = 0)
}
\arguments{
\item{x}{numeric}

\item{digits}{cantidad de dígitos del número entero que se quiere obtener.
Por defecto digits = 0.}
}
\value{
character
}
\description{
Transforma número a texto con separación
de miles con `.` y decimales con `,`.
}
\examples{

format_num(1000000)

}
