% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{frq_trunc}
\alias{frq_trunc}
\title{Funciones para revisar dados --------------------------------------------}
\usage{
frq_trunc(.data, .var = NULL, width = 50)
}
\description{
Funciones para revisar dados --------------------------------------------
}
