% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{is_email}
\alias{is_email}
\title{Validaci贸n de sintaxis de email}
\usage{
is_email(email)
}
\arguments{
\item{email}{vector character.}
}
\value{
vector logico
}
\description{
Validaci贸n de la sintaxis de correos electr贸nicos.
La expresi贸n regular utilizada para ello viene del siguiente post en [SO][https://stackoverflow.com/questions/201323/how-to-validate-an-email-address-using-a-regular-expression]
}
\examples{

is_email(c('a@a.com', 'a@a'))

}
