% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{kable_desuc}
\alias{kable_desuc}
\title{Función para ajustar estilo a tablas al momento de ser mostradas en un informe}
\usage{
kable_desuc(
  .data,
  digits = 1,
  row.names = NA,
  col.names = NA,
  align = NULL,
  caption = NULL,
  booktabs = TRUE,
  longtable = FALSE,
  escape = TRUE,
  font_size = 8,
  latex_options = c("hold_position"),
  ...
)
}
\arguments{
\item{.data}{Una data frame}

\item{digits}{\code{int} Número de decimales en la tabla (por defecto digits = 1)}

\item{row.names}{\code{Logical} Se incluye o no los nombres de las filas.}

\item{col.names}{\code{strings} Vector de texto para los nombres de las columnas.
(por defecto igual al colname de la tabla)}

\item{align}{\code{vector} Indicación para la alineación de las columnas. c('rll')}

\item{caption}{\code{string} Leyenda asociada a la tabla (por defecto sin leyenda)}

\item{booktabs}{\code{bolean} si usa estilo booktabs para la tabla (por defecto TRUE)}

\item{longtable}{\code{bolean} si usa estilo para tablas que cruzan dos páginas (por defecto FALSE)}

\item{escape}{\code{bolean} Si se normalizan los caracterres especiales en html o latex.
(por defecto TRUE)}

\item{font_size}{\code{ìnt} tamaño de letra en la tabla (por defecto 8)}

\item{latex_options}{\verb{vector de caracteres} donde se controla aspectos de latex.}

\item{...}{Atributos pasados a la función \code{kable_styling}}
}
\value{
Una kable con el formato DESUC
}
\description{
Función para ajustar estilo a tablas al momento de ser mostradas en un informe
}
