% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{pregunta_orden}
\alias{pregunta_orden}
\title{Orden de preguntas según valor de categoría}
\usage{
pregunta_orden(.x, .y, cat_orden)
}
\arguments{
\item{.x}{chr o factor. Variable con categorías de respuesta.}

\item{.y}{numeric. Valor de respuesta para cada categoría de respuesta}

\item{cat_orden}{string. Nombre de la categoría de respuesta de interés.}
}
\value{
numeric vector
}
\description{
función de ayuda para `forcats::fct_reorder2` con tal de ordenar distintas preguntas
según una catogoría de respuesta de interès `cat_orden`.
}
