% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_variables.R
\name{rec_cat_5a3}
\alias{rec_cat_5a3}
\title{Recodificaión de variables 5 a 3

Por defecto recodifica variables _likert_ de 5 niveles a 3 niveles.}
\usage{
rec_cat_5a3(variable, rec = "1:2 = 1; 3 = 2; 4:5 = 3; else = 9", labels = NULL)
}
\arguments{
\item{variable}{numeric, Valores de variable a recodificar.}

\item{rec}{string, por defecto recodifica 1:2 en 1, 3 en 2 y 4:5 en 3.
Todo lo demás lo deja como 9}

\item{labels}{vector string, etiquetas para las variables recodificadas.
Por defecto NULL.}
}
\value{
haven_labelled
}
\description{
Recodificaión de variables 5 a 3

Por defecto recodifica variables _likert_ de 5 niveles a 3 niveles.
}
\examples{

vect <- c(1, 4, 6, 99, NA)
rec_cat_5a3(vect, labels = c('alto' = 1))

}
