% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_variables.R
\name{rec_cat_7a3}
\alias{rec_cat_7a3}
\title{Recodificaión de variables 7 a 3}
\usage{
rec_cat_7a3(variable, rec = "1:4 = 1; 5 = 2; 6:7 = 3; else = 9", labels = NULL)
}
\arguments{
\item{variable}{numeric, Valores de variable a recodificar.}

\item{rec}{string, por defecto recodifica 1:4 en 1, 5 en 2 y 6:7 en 3.
Todo lo demás lo deja como 9}

\item{labels}{vector string, etiquetas para las variables recodificadas.
Por defecto NULL.}
}
\value{
haven_labelled
}
\description{
Por defecto recodifica variables de _notas_ de 7 niveles a 3 niveles.
}
\examples{

vect <- c(1, 4, 6, 99, NA)
rec_cat_7a3(vect, labels = c('alto' = 1))

}
