% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_orden.R
\name{region_orden}
\alias{region_orden}
\title{Orden de región}
\usage{
region_orden(reg, as.factor = TRUE)
}
\arguments{
\item{reg}{\code{num}. numérico con valores de 1 a 16}

\item{as.factor}{\code{logical}. Por defecto TRUE para crear factor ordenado geográficamente.
FALSE crea variable de clase \code{labelled}.}
}
\value{
ordered factor
}
\description{
Ordena un vector de números asociadas a las 16 regiones de Chile
según posición geográfica de norte a sur.
}
\examples{
region_orden(c(1, 13, 5, 15))

}
