% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_niveles.R
\name{rev_niveles}
\alias{rev_niveles}
\alias{rev_niveles.default}
\alias{rev_niveles.factor}
\alias{rev_niveles.haven_labelled}
\alias{rev_niveles.numeric}
\title{Invertir niveles}
\usage{
rev_niveles(x, ...)

\method{rev_niveles}{default}(x, ...)

\method{rev_niveles}{factor}(x, niveles_inv = NULL, ...)

\method{rev_niveles}{haven_labelled}(x, rango_inv = NULL, ...)

\method{rev_niveles}{numeric}(x, rango_inv = NULL, ...)
}
\arguments{
\item{x}{`vector`: Vector de datos que se quiere invertir}

\item{...}{}

\item{niveles_inv}{`chr`: Niveles de un factor que serán invertidos.
Por defecto, invierte todos los niveles.}

\item{rango_inv}{`num`: Vector de dos niveles en el que se señala el valor
del primero y último valor que será invertido. Se espera que esos dos valores
definan un rango.
Por defecto, invierte todos los niveles.}
}
\description{
Función para revertir niveles. Especialmente útil si se quiere revertir
sólo un rango de valores de una pregunta.
}
