% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut_informacion.R
\name{rut_información}
\alias{rut_información}
\title{Obtiene información para un RUT}
\usage{
rut_información(.rut)
}
\arguments{
\item{.rut}{string. rut sin puntos, con guión y dígito verificador.}
}
\value{
Una lista con datos asociados al RUT.
}
\description{
Obtiene información personal a partir de un RUT de Chile.
}
\examples{
rut_list = c(
"8714763-0",
"4606477-1",
"20283632-1"
) # example list of RUT values

l_ruts_info <- lapply(rut_list, rut_información)

l_ruts_info |> str(2)
}
