% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_date.R
\name{sg_date}
\alias{sg_date}
\title{Convierte una fecha para su uso en la API de SurveyToGo}
\usage{
sg_date(date, hora = "mañana")
}
\arguments{
\item{date}{Fecha en formato Date o POSIXct. Si se proporciona una fecha en formato POSIXct,
se mantendrá la hora proporcionada en el objeto.}

\item{hora}{Cadena de texto que indica si se debe tomar la 'mañana' (00:00:00) o la 'tarde'
(23:59:59) como hora de referencia para la fecha. Valor por defecto es 'mañana'.}
}
\value{
Un string con la fecha formateada en el estándar ISO 8601 extendido, incluyendo
milisegundos y la zona horaria correspondiente.
}
\description{
Esta función toma una fecha y la convierte al formato requerido por la API de SurveyToGo.
Si no se especifica una hora, se asume 'mañana' (00:00:00).
Si se especifica 'tarde', se asigna el último segundo del día (23:59:59).
}
\examples{
# Convertir una fecha en formato Date para la mañana
sg_date(as.Date('2024-08-26'))

# Convertir una fecha en formato Date para la tarde
sg_date(as.Date('2024-08-26'), hora = 'tarde')

# Convertir una fecha en formato POSIXct manteniendo la hora
sg_date(as.POSIXct('2024-08-26 14:35:00'))
}
