% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_get_data.R
\name{sg_get_data}
\alias{sg_get_data}
\title{SurveyToGo rest API: Get individual data}
\usage{
sg_get_data(
  api_operation = c("SimpleExport", "OperationData"),
  interview_ids,
  ...,
  api_key,
  user,
  pass,
  type = NA
)
}
\arguments{
\item{api_operation}{\code{chr} Bajar datos según operaciones "SimpleExport"
o "OperationData".}

\item{interview_ids}{\code{int} Vector con códigos de interviews de los que
se quiera bajar información.}

\item{...}{\code{list} pasa a query en sg_get}

\item{api_key}{\code{chr} REST API Key entregada por dooblo.}

\item{user}{\code{chr} nombre de usuario.}

\item{pass}{\code{chr} password de usuario.}

\item{type}{\code{chr} Formato en el que se obtendrán datos.
Por defecto 'application/json'. Puede ser 'text/xml' si se quiere xml o
'text/csv' si se desea un csv.}
}
\value{
data.frame
}
\description{
Función para bajar datos y respuestas considerando el límite de
99 casos por envío. Es un envoltorio de la función \code{sg_get()}.
}
\seealso{
sg_get
}
