% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{str_collapse}
\alias{str_collapse}
\title{Colapso de strings}
\usage{
str_collapse(.data, .var, collapse = ", ")
}
\arguments{
\item{.data}{tibble}

\item{.var}{variable name}

\item{collapse}{string, string con el que se unirán los textos. Por defecto
", ".}
}
\description{
Variable útil para colapsar en una sola casilla _list variables_ en un
tibble que quiera ser mostrad en una tabla
}
\examples{

df <- data.frame(x = c('a', 'b', 'c'))

str_collapse(df, x)

}
