% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{str_entre}
\alias{str_entre}
\title{Extrae string entre dos textos}
\usage{
str_entre(text, ini = "", fin = "")
}
\arguments{
\item{text}{string. Puede ser un named string.}

\item{ini}{string, desde donde se extrae el texto}

\item{fin}{string, hasta donde se extrae el texto}
}
\value{
string
}
\description{
Extrae string entre dos textos
}
