% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{str_entre_parentesis}
\alias{str_entre_parentesis}
\title{Función para extraer texto presente entre paréntesis.
Creado para trabajo con etiquetas de variables y extraer así el concepto
preguntado.}
\usage{
str_entre_parentesis(text)
}
\arguments{
\item{text}{texto}
}
\value{
string
}
\description{
Si no encuentra texto entre paréntesis, devuelve el texto original.
}
