% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_diff_grupo}
\alias{svy_diff_grupo}
\title{Comparación entre intervalos de confianza

Determina diferencias significativas según intervalos de confianza calculados desde
`svrvyr`.}
\usage{
svy_diff_grupo(.data, grupo = NA)
}
\arguments{
\item{.data}{data.frame con variables `\\*_upp` y `\\*_low`}

\item{grupo}{Variable por la que se verá la existencia de diferencias. Por defecto = NA.}
}
\value{
data.frame
}
\description{
Comparación entre intervalos de confianza

Determina diferencias significativas según intervalos de confianza calculados desde
`svrvyr`.
}
