% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_tabla_var_segmento}
\alias{svy_tabla_var_segmento}
\title{Tabla con intervalos de confianza}
\usage{
svy_tabla_var_segmento(
  .df,
  .var,
  .segmento = NULL,
  miss = NULL,
  vartype = c("ci", "se"),
  level = 0.95
)
}
\arguments{
\item{.df}{data frame con diseño complejo}

\item{.var}{Variable en la que interesa comparar categorías de respuesta.}

\item{.segmento}{Segmentos de interés para ver diferencias en categorías
de variable \code{.var}. Por defecto NULL}

\item{miss}{chr vector. Categorías a excluir en el denominador de \code{prop_val} y \code{mean}.}

\item{vartype}{chr vector. Tipo de error para estimaciones (\code{ci}, \code{se}, etc.).}

\item{level}{double. Nivel de significancia para intervalos de confianza}
}
\value{
tibble
}
\description{
Devuelve tabla de frecuencias con intervalos de confianza para un nivel
\code{level}de significancia entre las categorías de respuesta de la
variable \code{.var}.
}
