% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_tabla_vars_segmentos}
\alias{svy_tabla_vars_segmentos}
\title{Tabla con intervalos de confianza}
\usage{
svy_tabla_vars_segmentos(
  .df,
  .vars,
  .segmentos = NULL,
  miss = NULL,
  vartype = c("ci", "se"),
  level = 0.95
)
}
\arguments{
\item{.df}{\code{tbl_svy} data.frame con diseño de encuesta.}

\item{.vars}{c(). Variables de interés respecto.}

\item{.segmentos}{c(). Lista de variables por las que se quiere segmentar \code{.vars}.}

\item{miss}{chr vector. Categorías a excluir en el denominador de \code{prop_val} y \code{mean}.}

\item{vartype}{chr vector. Tipo de error para estimaciones (\code{ci}, \code{se}, etc.).}

\item{level}{double. Nivel de confianza para intervalos.}
}
\value{
tibble
}
\description{
Devuelve tabla de frecuencias con intervalos de confianza para un nivel
\code{level}de significancia entre las categorías de respuesta de la
variable \code{.vars}.

A diferencia de svy_tabla_var_segmentos, esta funcion puede procesar varias variables
y segmentos a la vez.
}
