% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_categoria}
\alias{tabla_categoria}
\title{Tablas de resultados ----------------------------------------------------}
\usage{
tabla_categoria(.df, .var, .wt = NULL, miss = NULL)
}
\arguments{
\item{.df}{tibble. datos tidy.}

\item{.var}{chr. Variable de la que se sabrá su distribución. Solo una variable.}

\item{.wt}{chr. Variable de ponderación. Solo una variable.}

\item{miss}{chr vector. Nombre de niveles de categorías de `.var` que deben
ser considerados missing.}
}
\description{
Produce una tabla de distribución por categorías para una variable.
}
\details{
Generación de una data.frame con el número de casos y proporción de las distintas
variables de segmentos que se agregen en `...`.
}
