% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_categorias}
\alias{tabla_categorias}
\title{Tabla de categorías}
\usage{
tabla_categorias(.df, ..., .wt = NULL, miss = NULL)
}
\arguments{
\item{.df}{data.frame. Base de datos.}

\item{...}{Preguntas de las que se quiere saber su proporcion. Se puede utilizar
\code{tidyselect} para facilitar la selección de varias columnas.}

\item{.wt}{name. Nombre de columna con ponderador o expansor de los datos. Por defecto es NULL,
sin expansor.}

\item{miss}{chr vector. Categorías de respuesta que deben excluirse del cálculo de \code{prop_val}.}
}
\value{
tibble
}
\description{
Porcentaje de respuesta de categorías de varias variables.
Principalmente para mostrar la distribución de casos de variables de segmentación posteriores.
}
