% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{tabla_columnas}
\alias{tabla_columnas}
\title{Divide un data.frame en columnas

Toma una tabla y la divide y pega en sucesicas columnas para imprimir en menor espacio.}
\usage{
tabla_columnas(data, ncols = 2)
}
\arguments{
\item{data}{tibble}

\item{ncols}{integer. Número de columnas en las que se quiere dividir la data original.
Por defecto ncols = 2.}
}
\value{
tibble
}
\description{
Divide un data.frame en columnas

Toma una tabla y la divide y pega en sucesicas columnas para imprimir en menor espacio.
}
