% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_prop}
\alias{tabla_prop}
\title{Cálculo de proporciones}
\usage{
tabla_prop(.df, .segmento)
}
\arguments{
\item{.df}{tibble.}

\item{.segmento}{chr.}
}
\value{
tibble
}
\description{
Cálculo de proporciones
}
\examples{
df <- tibble::tibble(var = c("a", "b"),
            casos = c(30, 70))

desuctools:::tabla_prop(df, .segmento = "var")

}
