% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_prop_val}
\alias{tabla_prop_val}
\title{Calculo de proporción para categorías válidas.}
\usage{
tabla_prop_val(.df, by = NULL, miss)
}
\arguments{
\item{.df}{tibble}

\item{by}{chr vector. variables respecto de la que se calculará el porcentaje.}

\item{miss}{chr vector. Categorías que se deben considerar como missing o inválidas.}
}
\value{
tibble
}
\description{
Calculo de proporción para categorías válidas.
}
\examples{
df <- tibble(pregunta_cat = c("a", "b", "x"),
             casos = c(30, 30, 40))

tabla_prop_val(df, miss = "x")

tabla_prop_val(df, by = "pregunta_cat", miss = "x")

}
