% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_total_prop}
\alias{tabla_total_prop}
\title{Agrega dato total por segmento.}
\usage{
tabla_total_prop(.df)
}
\arguments{
\item{.df}{tibble}
}
\value{
tibble
}
\description{
Es una ayuda para tabla_vars_segmentos.
}
