% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_var_segmento}
\alias{tabla_var_segmento}
\title{Tabla de respuestas de una variable para cada categoría de un segmento.}
\usage{
tabla_var_segmento(
  .df,
  .var,
  .segmento = NULL,
  .wt = NULL,
  miss = NULL,
  total = FALSE
)
}
\arguments{
\item{.df}{tibble.}

\item{.var}{chr.}

\item{.segmento}{chr.}

\item{.wt}{chr.}

\item{miss}{chr vector.}

\item{total}{logic}
}
\value{
tibble
}
\description{
Resuelve el caso particular para tabla_vars_segmentos.
}
