% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames_to_label.R
\name{varnames_to_label}
\alias{varnames_to_label}
\title{varnames_to_label}
\usage{
varnames_to_label(
  .df,
  pattern_detect,
  pattern_extract = NULL,
  var_prefix = NULL
)
}
\arguments{
\item{.df}{`data.frame` Un data.frame}

\item{pattern_detect}{`chr` patrón de regex con el que se extrae el nombre breve
de la variable. Este es el que será utilizado posteriormente como
nombre final de la variable en la base}

\item{pattern_extract}{`chr` patrón de regex opcional.
Por defecto es NULL e implica que se usará el mismo patrón que el dispuesto
en `pattern_detect`.}

\item{var_prefix}{`chr` opcional. Prefijo para el nombre de las nuevas variables
que se extraigan.}
}
\value{
data.frame
}
\description{
Toma nombre de variables y los separa en un nombre breve
      y el resto de la etiqueta la integra en el atributo `label` de la
      variable.
      Además cambia la clase de las variables modificadas a `haven_labelled`
}
\examples{
df <- data.frame(`first variable 1` = 1,
                 `second variable 2` = 2,
                 var_1 = 3L,
                 check.names = FALSE)

varnames_to_label(df, pattern_detect = ' \\\\d$',
                  var_prefix = 'p')

varnames_to_label(df,
                  pattern_detect = ' \\\\d$',
                  pattern_extract = '\\\\d$',
                  var_prefix = 'p')

}
