% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alchemer_functions.R
\name{alch_read_spss}
\alias{alch_read_spss}
\title{read Alchemer SPSS export}
\usage{
alch_read_spss(url)
}
\arguments{
\item{url}{`chr` Distribution link del reporte de base de datos en SPSS.}
}
\value{
tibble
}
\description{
Lee el archivo .sav a partir del _distribution link_ de un reporte de
exportación de una encuesta programada en Alchemer.
https://help.alchemer.com/help/spss

Compatibility
- Comments are not available in SPSS exports.
- The Conjoint question is not available in SPSS exports.
- "Other, Write-In" rows in grid questions as not available in SPSS exports.
- The Time Started field is not includes as part of the SPSS export.
}
