% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_nse_desuc.R
\name{calculo_nse}
\alias{calculo_nse}
\title{Funciones de trabajo con datos ------------------------------------------}
\usage{
calculo_nse(.data, .ocupacion_jh, .educacion_jh, .bienes_var = NULL,
  append = TRUE)
}
\description{
Cálculo de Nivel socioeconómico según formula DESUC utilizando
ocupación, educación y bienes.
}
