% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{digitos_entero}
\alias{digitos_entero}
\title{Title Número de dígitos enteros.}
\usage{
digitos_entero(x, digits = 2)
}
\arguments{
\item{x}{vector numerico.}

\item{digits}{cantidad de dígitos del número entero que se quiere obtener.
Por defecto digits = 2.}
}
\value{
vector numerico.
}
\description{
Número de dígitos enteros para valores con decimales. Diseñada para casos en que
longitud y latitud son capturados como digitos sin decimal y necesitan ser reescalados.
}
