% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edad_rut.R
\name{edad_rut}
\alias{edad_rut}
\title{Edad_rut}
\usage{
edad_rut(.rut, fecha_referencia)
}
\arguments{
\item{.rut}{Vector numérico con el rut (sin dígito verificador).}

\item{fecha_referencia}{Vector que contiene la fecha (yy-mm-dd) que determina la edad.
Pueden ser años pasados, el año actual o años futuros.}
}
\value{
integer
}
\description{
Función para calcular la edad según el rut de la persona.
}
\examples{

# Importante: el rut no debe contar con el dígito verificador
x <- 20117419
fecha <- "2024-01-31"

edad_rut(.rut = x,
         fecha_referencia = fecha)



}
