% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_es_educacion_adulto}
\alias{edu_es_educacion_adulto}
\title{Detecta códigos de enseñanza para adultos}
\usage{
edu_es_educacion_adulto(x)
}
\arguments{
\item{x}{Un vector numérico o de cadenas.}
}
\value{
Un vector lógico indicando si es un código de enseñanza para adultos.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Detecta niveles de enseñanza para adultos ya que su segunda cifra significativa
es igual al dígito \code{6}.
}
\seealso{
\link{https://centroestudios.mineduc.cl/datos-abiertos/}
}
