% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_cod_ense}
\alias{edu_fct_cod_ense}
\title{Factor de niveles de ense\u00f1anza}
\usage{
edu_fct_cod_ense(x)
}
\arguments{
\item{x}{Un vector de c\u00f3digos, t\u00edpicamente variable `COD_ENSE` o `COD_ENS_X`.}
}
\value{
A factor with 44 levels, with an attached `"label"` attribute.
}
\description{
Recodificaci\u00f3n para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Niveles de ense\u00f1anza
}
\seealso{
[https://centroestudios.mineduc.cl/datos-abiertos/]
}
