% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_cod_ense}
\alias{edu_fct_cod_ense}
\title{Factor de niveles de enseñanza}
\usage{
edu_fct_cod_ense(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{COD_ENSE} o \code{COD_ENS_X}.}
}
\value{
A factor with 44 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Niveles de enseñanza
}
\seealso{
\link{https://centroestudios.mineduc.cl/datos-abiertos/}
}
