% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_cod_ense2}
\alias{edu_fct_cod_ense2}
\title{Factor de niveles de enseñanza agrupado}
\usage{
edu_fct_cod_ense2(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{COD_ENSE2}.}
}
\value{
A factor with 8 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Niveles de enseñanza agrupados
}
\seealso{
\link{https://centroestudios.mineduc.cl/datos-abiertos/}
}
