% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_cod_ense3}
\alias{edu_fct_cod_ense3}
\title{Factor de niveles de enseñanza agrupado}
\usage{
edu_fct_cod_ense3(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{COD_ENSE3}.}
}
\value{
A factor with 7 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Niveles de enseñanza con educación especial reasignado.
}
\seealso{
\url{https://centroestudios.mineduc.cl/datos-abiertos/}
}
