% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_educacion_especial}
\alias{edu_fct_educacion_especial}
\title{Factor para distingir Educaci\u00f3n Especial}
\usage{
edu_fct_educacion_especial(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A factor with 3 levels, with an attached `"label"` attribute.
}
\description{
Recodificaci\u00f3n para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Diferencia Educaci\u00f3n Especial en tres niveles.
es igual al d\u00edgito `6`.

Los cod_educ relacioandos son:

- 211 Educaci\u00f3n Especial Discapacidad Auditiva
- 212 Educaci\u00f3n Especial Discapacidad Intelectual
- 213 Educaci\u00f3n Especial Discapacidad Visual
- 214 Educaci\u00f3n Especial Trastornos Espec\u00edficos del Lenguaje
- 215 Educaci\u00f3n Especial Trastornos Motores
- 216 Educaci\u00f3n Especial Autismo
- 217 Educaci\u00f3n Especial Discapacidad Graves Alteraciones en la Capacidad de Relaci\u00f3n y Comunicaci\u00f3n
- 218 Educaci\u00f3n Especial Discapacidad M\u00faltiple
- 219 Educaci\u00f3n Especial Sordoceguera
- 299 Opci\u00f3n 4 Programa Integraci\u00f3n Escolar
}
\examples{
edu_fct_educacion_especial(c(211, 211, 216, 214, 110)) |>
table()

}
\seealso{
[https://centroestudios.mineduc.cl/datos-abiertos/]
}
