% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_educacion_especial}
\alias{edu_fct_educacion_especial}
\title{Factor para distingir Educación Especial}
\usage{
edu_fct_educacion_especial(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A factor with 3 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Diferencia Educación Especial en tres niveles.
es igual al dígito \code{6}.

Los cod_educ relacioandos son:
\itemize{
\item \code{211}: Educación Especial Discapacidad Auditiva
\item \code{212}: Educación Especial Discapacidad Intelectual
\item \code{213}: Educación Especial Discapacidad Visual
\item \code{214}: Educación Especial Trastornos Específicos del Lenguaje
\item \code{215}: Educación Especial Trastornos Motores
\item \code{216}: Educación Especial Autismo
\item \code{217}: Educación Especial Discapacidad Graves Alteraciones en la Capacidad de Relación y Comunicación
\item \code{218}: Educación Especial Discapacidad Múltiple
\item \code{219}: Educación Especial Sordoceguera
\item \code{299}: Opción 4 Programa Integración Escolar
}
}
\examples{
edu_fct_educacion_especial(c(211, 211, 216, 214, 110)) |>
table()

}
\seealso{
\url{https://centroestudios.mineduc.cl/datos-abiertos/}
}
