% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_educacion_especial}
\alias{edu_fct_educacion_especial}
\title{Factor para distingir Educación Especial}
\usage{
edu_fct_educacion_especial(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A factor with 3 levels, with an attached `"label"` attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Diferencia Educación Especial en tres niveles.
es igual al dígito `6`.

Los cod_educ relacioandos son:

- 211 Educación Especial Discapacidad Auditiva
- 212 Educación Especial Discapacidad Intelectual
- 213 Educación Especial Discapacidad Visual
- 214 Educación Especial Trastornos Específicos del Lenguaje
- 215 Educación Especial Trastornos Motores
- 216 Educación Especial Autismo
- 217 Educación Especial Discapacidad Graves Alteraciones en la Capacidad de Relación y Comunicación
- 218 Educación Especial Discapacidad Múltiple
- 219 Educación Especial Sordoceguera
- 299 Opción 4 Programa Integración Escolar
}
\examples{
edu_fct_educacion_especial(c(211, 211, 216, 214, 110)) |>
table()

}
\seealso{
[https://centroestudios.mineduc.cl/datos-abiertos/]
}
