% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_ens}
\alias{edu_fct_ens}
\title{Factor de c\u00f3digo de ense\u00f1anza en 13 niveles}
\usage{
edu_fct_ens(x)
}
\arguments{
\item{x}{Un vector de c\u00f3digos, t\u00edpicamente variable `COD_DEPE2`.}
}
\value{
A factor with 13 levels, with an attached `"label"` attribute.
}
\description{
Recodificaci\u00f3n para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Recodificaci\u00f3n del C\u00f3digo de Ense\u00f1anza en 13 niveles.
}
\seealso{
[https://centroestudios.mineduc.cl/datos-abiertos/]
}
