% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_educacion.R
\name{edu_fct_rural_rbd}
\alias{edu_fct_rural_rbd}
\title{Factor de ruralidad}
\usage{
edu_fct_rural_rbd(x)
}
\arguments{
\item{x}{Un vector de códigos, típicamente variable \code{RURAL_RBD}.}
}
\value{
A factor with 2 levels, with an attached \code{"label"} attribute.
}
\description{
Recodificación para bases de datos disponibles en el Centro de Estudios MINEDUC, Chile
Índice de ruralidad del establecimiento.
}
\seealso{
\link{https://centroestudios.mineduc.cl/datos-abiertos/}
}
