% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{frq_trunc}
\alias{frq_trunc}
\title{Frecuencia de variable con etiqueta truncada}
\usage{
frq_trunc(..., width = 50L, ellipsis = "...")
}
\arguments{
\item{...}{Una data frame o vector según lo requerido por `sjmisc::frq`.}

\item{width}{numeric. Por defecto = 50. Largo del texto de la etiqueta de la variable.}

\item{ellipsis}{string. Por defecto = '...'.}
}
\value{
Una kable con el formato DESUC
}
\description{
Se trunca el texto de la etiqueta de la variable a un largo preestablecido para
una mejor visión de la tabla resultante.
}
