% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos_output.R
\name{gg_bar_3_niveles_stack}
\alias{gg_bar_3_niveles_stack}
\title{Barras 3 niveles positivo neutro negativo

Gr\\u00e1fico de barras dise\\u00f1ado para comparar categor\\u00eda positiva, negativa y neutra.}
\usage{
gg_bar_3_niveles_stack(
  .df,
  x = pregunta_lab,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  missing = NULL,
  text_size = 3,
  flip = TRUE,
  colour_neg_neu_pos = c("#C00001", "#FFC000", "#20497D"),
  y_prop = prop,
  y_na = 1.1,
  x_str_entre_ini = "",
  x_str_entre_fin = "",
  x_str_width = 50,
  colour_na = "grey20",
  font_family = "Calibri"
)
}
\arguments{
\item{x}{`quo` Nombre de variable a utilizar en eje X.}

\item{title}{`chr` T\\u00edtulo del gr\\u00e1fico}

\item{subtitle}{`chr` Subt\\u00edtulo del gr\\u00e1fico}

\item{caption}{`chr` Caption del gr\\u00e1fico}

\item{missing}{`chr` vector con categor\\u00edas de respuesta consideradas 'missing'}

\item{text_size}{`num` tama\\u00f1o de letra}

\item{flip}{`logical` TRUE gira los ejes.}

\item{font_family}{}

\item{.data}{`data.frame` Debe contener variables `pregunta_lab` y `pregunta_cat`.
Funciona bien a partir de data.frame de resultado de funci\\u00f3n `tabla_vars_segmentos`.}
}
\value{
ggplot
}
\description{
Barras 3 niveles positivo neutro negativo

Gr\\u00e1fico de barras dise\\u00f1ado para comparar categor\\u00eda positiva, negativa y neutra.
}
\examples{
df_chart <- data.frame(pregunta_lab = c(rep('a', 4), rep('b', 4)),
                       x_other = c(rep('x', 4), rep('y', 4)),
                       prop = c(-0.1, 0.3, 0.4, 0.1, -0.3, 0.1, 0.4, 0.05),
                       pregunta_cat = factor(rep(c('bajo', 'medio', 'alto', 'ns'), 2),
                                             levels = c('bajo', 'medio', 'alto', 'ns')))

gg_bar_3_niveles_stack(df_chart,
                       missing = 'ns',
                       title = 'Grafico de prueba')

}
