% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_print.R
\name{kable_desuc}
\alias{kable_desuc}
\title{Función para ajustar estilo a tablas al momento de ser mostradas en un informe}
\usage{
kable_desuc(.data, caption = NULL, ...)
}
\arguments{
\item{.data}{Una data frame}

\item{caption}{Leyenda asociada a la tabla}

\item{...}{Atributos pasados a la función `kable_styling`}
}
\value{
Una kable con el formato DESUC
}
\description{
Función para ajustar estilo a tablas al momento de ser mostradas en un informe
}
