% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_labelled.R
\name{rec_labelled}
\alias{rec_labelled}
\title{Recodificación de variables labelled}
\usage{
rec_labelled(.vec, ls_niveles)
}
\arguments{
\item{.vec}{vector/variable a recodificar}

\item{ls_niveles}{lista con recodificaciones.
Debe cumplir un formato específico.}
}
\value{
vector labelled recodificado
}
\description{
Función para recodificar variables labelled utilizando una lista
}
\examples{
# Crear un vector etiquetado
x <- labelled::labelled(
  c(1, 2, 2, 2, 3, 4, 5, 5, 1, NA),
  labels = c("Muy de acuerdo" = 1,
             "De acuerdo" = 2,
             "Ni de acuerdo ni en desacuerdo" = 3,
             "En desacuerdo" = 4,
             "Muy en desacuerdo" = 5)
)

# Definir la lista de recodificaciones
# IMPORTANTE: el formato es "Val_final-Etiqueta = Val_inicial"
# Donde:
#    Val_final: serán los valores finales de las etiquetas.
#    Etiqueta: etiquetas de los valores anteriormente definidos.
#    Val_inicial: categorías que se consideran para la recodificación (numéricos)
ls_rec <- list(
  "1-Muy de acuerdo + De acuerdo" = 1:2,
  "2-Ni de acuerdo ni en desacuerdo" = 3,
  "3-En desacuerdo + Muy en desacuerdo" = 4:5,
  "99-NA" = NA
)

# Aplicar rec_labelled al vector etiquetado
resultado_recodificado <- rec_labelled(.vec = x, ls_niveles = ls_rec)
resultado_recodificado



}
