% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_nse_aim2023.R
\name{rec_nse_aim2023}
\alias{rec_nse_aim2023}
\title{Clasificación NSE AIM 2023}
\source{
https://aimchile.cl/wp-content/uploads/2025/06/Actualizacion-y-Manual-GSE-AIM-2023.pdf
}
\usage{
rec_nse_aim2023(.edu_jh, .ocu_jh, .tramo_ingreso, factor = FALSE)
}
\arguments{
\item{.edu_jh}{Integer. Vector de códigos de educación (P1 AIM: 1-5).}

\item{.ocu_jh}{Integer. Vector de códigos de ocupación (P2 AIM: 1-6).}

\item{.tramo_ingreso}{Integer. Vector de tramo de ingreso (1-7, 99 = NS/NR).}

\item{factor}{logical. Si es \code{TRUE}, retorna factor ordenado; si es \code{FALSE}, retorna labelled.}
}
\value{
labelled. Vector con la clasificación GSE.
}
\description{
Clasifica el nivel socioeconómico (GSE) segun la
actualización AIM 2023, usando:
\itemize{
\item edu_jh: educación jefe(a) de hogar (códigos edu_jh: 1-5)
\item ocu_jh: ocupación jefe(a) de hogar (códigos ocu_jh: 1-6)
\item residentes: numero de personas residentes en el hogar (>=1)
\item tramo_ingreso: Tramos de ingreso autónomo per cápita del hogar (1-7, 99 = NS/NR)
}

Devuelve un vector de caracteres con el GSE: E, D, C3, C2, C1b, C1a, AB.

Preguntas:

Educación

\code{edu_jh}: ¿Cuál es el nivel educacional alcanzado (último año aprobado) por el principal sostenedor del hogar?
\itemize{
\item 1 a 1 Sin estudios formales
\item 2 a 1 Básica incompleta / primaria o preparatoria incompleta
\item 3 a 2 Básica completa / primaria o preparatoria completa
\item 4 a 3 Media científico humanista o media técnico profesional incompleta / humanidades incompleta
\item 5 a 3 Media científico humanista o media técnico profesional completa / humanidades completa
\item 6 a 4 Instituto técnico (CFT) o instituto profesional incompleto (carreras 1 a 3 años)
\item 7 a 4 Instituto técnico (CFT) o instituto profesional completo (carreras 1 a 3 años) / hasta suboficial de FF.AA./Carabineros
\item 8 a 4 Universitaria incompleta (carreras de 4 o mas años)
\item 9 a 5 Universitaria completa (carreras de 4 o mas años) / oficial de FF.AA./Carabineros
\item 10 a 5 Postgrado (Postítulo, Máster, Magíster, Doctor)
}

\code{ocu_jh} ¿Cuál de las siguientes ocupaciónes corresponde al trabajo del principal sostenedor del hogar?
(SI EL PRINCIPAL SOSTENEDOR DEL HOGAR ESTA CESANTE O ES JUBILADO, PREGUNTAR POR LA ULTIMA OCUPACION REMUNERADA QUE TUVO)
(SI EL PRINCIPAL SOSTENEDOR TIENE MAS DE 1 TRABAJO, DEBE REGISTRARSE EL DE MAYOR INGRESO)
\itemize{
\item 1 a 1 Trabajadores no calificados en ventas y servicios, peones agropecuarios, forestales, construccion, etc.
\item 2 a 2 Obreros, operarios y artesanos de artes mecánicas y de otros oficios
\item 3 a 3 Trabajadores de los servicios y vendedores de comercio y mercados
\item 4 a 3 Agricultores y trabajadores calificados agropecuarios y pesqueros
\item 5 a 3 Operadores de instalaciones y máquinas y montadores / conductores de vehículos
\item 6 a 3 Otros grupos no identificados (incluye rentistas, incapacitados, etc.)
\item 7 a 4 Empleados de oficina públicos y privados
\item 8 a 5 Técnicos y profesionales de nivel medio (incluye hasta suboficiales FF.AA./Carabineros)
Alto ejecutivo (gerente general o gerente de area o sector) de empresa privadas o públicas. Director o dueño de
grandes empresas. Alto directivo del poder ejecutivo, de los cuerpos legislativos y de la administracion publica
(incluye oficiales de FF.AA./carabineros)
\item 9 a 6 Profesionales, científicos e intelectuales
}
}
\examples{
# personas <- data.frame(
#   edu_jh        = c(3L, 5L, 2L, 4L),
#   ocu_jh        = c(1L, 6L, 4L, 3L),
#   residentes    = c(4L, 5L, 2L, 9L),
#   tramo_ingreso = c(4L, 7L, 3L, 2L)
# )
# personas$gse <- rec_nse_aim2023(
#   edu_jh        = personas$edu_jh,
#   ocu_jh        = personas$ocu_jh,
#   residentes    = personas$residentes,
#   tramo_ingreso = personas$tramo_ingreso
# )
}
