% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-regiones_y_comunas.R
\docType{data}
\name{regiones_y_comunas}
\alias{regiones_y_comunas}
\title{Datos de regiones y comunas de Chile}
\format{

}
\usage{
regiones_y_comunas
}
\description{
Tabla que contiene nombres y códigos de regiones y comunas de Chile,
además de algunas segmentación es utilizadas habitualmente
}
\keyword{datasets}
