% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpus_rae.R, R/data-regiones_y_comunas.R
\docType{data}
\name{regiones_y_comunas}
\alias{regiones_y_comunas}
\title{Corpus de Referencia del Español Actual (CREA) escrito}
\format{
Un data.frame con 3 variables


}
\source{
\url{https://www.rae.es/recursos/banco-de-datos/crea-escrito}

\describe{
  \item{Orden}{formas ortográficas registradas en el CREA}
  \item{Frec.absoluta}{frecuencia absoluta de la palabra}
  \item{Frec.normalizada}{frecuencia normalizada de la palabra}
}
}
\usage{
regiones_y_comunas

regiones_y_comunas
}
\description{
Un corpus de referencia ha de ser lo suficientemente extenso para representar
todas las variedades relevantes de una lengua. En la versión 3.2, cerrada en junio de 2008
y consultable en la actualidad, la parte escrita del CREA cuenta con casi 140 000 documentos
y algo más de 154 millones de formas procedentes de textos de todos los países
hispánicos y producidos entre 1975 y 2004.

Tabla que contiene nombres y códigos de regiones y comunas de Chile,
además de algunas segmentación es utilizadas habitualmente.
}
\keyword{datasets}
