% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpus_rae.R, R/data-regiones_y_comunas.R
\docType{data}
\name{regiones_y_comunas}
\alias{regiones_y_comunas}
\title{Corpus de Referencia del Español Actual (CREA) escrito}
\format{
Un data.frame con 3 variables

Un data.frame con 346 filas \code{comunas} y 19 variables.
\describe{
  \item{id}{orden de comunas siguiendo geografía y código comunal}
  \item{region_orden}{numeración de regiones de norte a sur}
  \item{comuna}{código de comuna}
  \item{comuna18}{código de comuna luego de la creación de la región de Ñuble en 2018}
  \item{region}{número de región}
  \item{provincia}{código de provincia luego de la creación de la región de Ñuble}
  \item{region_nom}{nombre oficial de la región}
  \item{region_nom_small}{nombre breve de la región}
  \item{comuna_nom}{nombre de la comuna}
  \item{region_capital}{TRUE si comuna es capital regional}
  \item{comuna_extrema}{1 si comuna es extrema para INE, 2 si es extrema para DESUC}
  \item{gran_santiago}{TRUE si comuna pertenece al Gran Santiago}
  \item{gran_santiago_zonas}{nombre de zonas del Gran Santiago}
  \item{educ_sup_com}{\% hogares con algún miembro con educación superior}
  \item{hacinamiento}{\% hogares con hacinamiento}
  \item{allegamiento_int}{\% hogares con allegamiento interno}
  \item{carente_ser_bas}{\% hogares carentes de servicios básicos}
  \item{carente_entorno}{\% hogares carentes de características beneficiosas en el entorno}
  \item{decil}{promedio de decir de ingresos de los hogares}
}
}
\source{
\url{https://www.rae.es/recursos/banco-de-datos/crea-escrito}

\describe{
  \item{Orden}{formas ortográficas registradas en el CREA}
  \item{Frec.absoluta}{frecuencia absoluta de la palabra}
  \item{Frec.normalizada}{frecuencia normalizada de la palabra}
}

<http://www.ine.cl/>
}
\usage{
regiones_y_comunas

regiones_y_comunas
}
\description{
Un corpus de referencia ha de ser lo suficientemente extenso para representar
todas las variedades relevantes de una lengua. En la versión 3.2, cerrada en junio de 2008
y consultable en la actualidad, la parte escrita del CREA cuenta con casi 140 000 documentos
y algo más de 154 millones de formas procedentes de textos de todos los países
hispánicos y producidos entre 1975 y 2004.

Tabla que contiene nombres y códigos de regiones y comunas de Chile,
además de algunas segmentación es utilizadas habitualmente.
}
\keyword{datasets}
