% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytogo_rest_api.R
\name{sg_get}
\alias{sg_get}
\title{SurveyToGo rest API}
\usage{
sg_get(api_operation, query, api_key, user, pass, type = "application/json")
}
\arguments{
\item{api_operation}{`chr`}

\item{query}{`list` listado de variables que serán pasadas a la api_operation
que se elija.}

\item{api_key}{`chr` REST API Key entregada por dooblo.}

\item{user}{`chr` nombre de usuario.}

\item{pass}{`chr` password de usuario.}

\item{type}{`chr` Formato en el que se obtendrán datos.
Por defecto 'application/json'. Puede ser 'text/xml' si se quiere xml.}
}
\value{
json
}
\description{
Función para acceder a datos del sistema SurveyToGo utilizando
       su API REST.
}
\references{
https://support.dooblo.net/hc/en-us/articles/208294645-How-To-Use-The-SurveyToGo-REST-API
}
