% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_get_data.R
\name{sg_get_data}
\alias{sg_get_data}
\title{SurveyToGo rest API: Get individual data}
\usage{
sg_get_data(
  api_operation = c("SimpleExport", "OperationData"),
  interview_ids,
  ...,
  api_key,
  user,
  pass,
  type = NA
)
}
\arguments{
\item{api_operation}{`chr` Bajar datos según operaciones "SimpleExport"
o "OperationData".}

\item{interview_ids}{`int` Vector con códigos de interviews de los que
se quiera bajar información.}

\item{...}{`list` pasa a query en sg_get}

\item{api_key}{`chr` REST API Key entregada por dooblo.}

\item{user}{`chr` nombre de usuario.}

\item{pass}{`chr` password de usuario.}

\item{type}{`chr` Formato en el que se obtendrán datos.
Por defecto 'application/json'. Puede ser 'text/xml' si se quiere xml o
'text/csv' si se desea un csv.}
}
\value{
data.frame
}
\description{
Función para bajar datos y respuestas considerando el límite de
      99 casos por envío. Es un envoltorio de la función `sg_get()`.
}
\seealso{
sg_get
}
