% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_variables.R
\name{shift_missing}
\alias{shift_missing}
\title{Corregir missings en preguntas múltiples}
\usage{
shift_missing(.data, .x, .y = NULL, missing = c(77L, 88L, 99L))
}
\arguments{
\item{.data}{Una data frame}

\item{.x}{nombre de la variable la primera variable}

\item{.y}{nombre de la variable la primera variable}

\item{missing}{vector con valores considerados no válidos (por defecto \code{77, 88, 99})}
}
\value{
Una data frame con los valores de .x y .y corregidos
}
\description{
Función para corregir problemas de no respuesta en preguntas múltiples
}
\details{
Funciones de trabajo con datos ------------------------------------------
}
