% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_tabla_var_segmento}
\alias{svy_tabla_var_segmento}
\title{Tabla con intervalos de confianza}
\usage{
svy_tabla_var_segmento(
  .df,
  .var,
  .segmento = NULL,
  miss = NULL,
  vartype = c("ci", "se"),
  level = 0.95
)
}
\arguments{
\item{.var}{Variable en la que interesa comparar categorías de respuesta.}

\item{.segmento}{Segmentos de interés para ver diferencias en categorías
de variable `.var`. Por defecto NULL}

\item{level}{double. Nivel de significancia para intervalos de confianza}

\item{.data}{data frame con diseño complejo}

\item{na.rm}{boolean. Considera o no missings, por defecto FALSE.}
}
\value{
tibble
}
\description{
Devuelve tabla de frecuencias con intervalos de confianza para un nivel
`level`de significancia entre las categorías de respuesta de la
variable `.var`.
}
