% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_tabla_var_segmento}
\alias{svy_tabla_var_segmento}
\title{Tabla con intervalos de confianza

Devuelve tabla de frecuencias con intervalos de confianza para un nivel
`level`de significancia entre las categorías de respuesta de la
variable `.var`.}
\usage{
svy_tabla_var_segmento(.data, .var, .segmento = NULL, na.rm = TRUE,
  level = 0.95)
}
\arguments{
\item{.data}{data frame con diseño complejo}

\item{.var}{Variable en la que interesa comparar categorías de respuesta.}

\item{na.rm}{boolean. Considera o no missings, por defecto FALSE.}

\item{level}{double. Nivel de significancia para intervalos de confianza}

\item{.segmentos}{Segmentos de interés para ver diferencias en categorías
de variable `.var`. Por defecto NULL}
}
\value{
data.frame
}
\description{
Tabla con intervalos de confianza

Devuelve tabla de frecuencias con intervalos de confianza para un nivel
`level`de significancia entre las categorías de respuesta de la
variable `.var`.
}
