% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_survey.R
\name{svy_tabla_vars_segmentos}
\alias{svy_tabla_vars_segmentos}
\title{Tabla con intervalos de confianza}
\usage{
svy_tabla_vars_segmentos(.data, .vars, .segmentos = NULL, ...)
}
\arguments{
\item{.data}{`tbl_svy` data.frame con diseño de encuesta.}

\item{.vars}{c(). Variables de interés respecto.}

\item{.segmentos}{c(). Lista de variables por las que se quiere segmentar `.vars`.}

\item{...}{atributos que se pasan a funcion `svy_tabla_var_segmento`.}
}
\value{
data.frame
}
\description{
Devuelve tabla de frecuencias con intervalos de confianza para un nivel
`level`de significancia entre las categorías de respuesta de la
variable `.vars`.

A diferencia de svy_tabla_var_segmentos, esta funcion puede procesar varias variables
y segmentos a la vez.
}
