% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tablas.R
\name{tabla_categorias}
\alias{tabla_categorias}
\title{Tabla de categorías

Porcentaje de respuesta de categorías de varias variables.
Principalmente para mostrar la distribución de casos de variables de segmetnación posteriores.}
\usage{
tabla_categorias(.data, ..., .wt = NULL)
}
\arguments{
\item{.data}{data frame. Base de datos.}

\item{...}{Preguntas de las que se quiere saber su proporcion. Se puede utilizar
`tidyselect` para facilitar la selección de varias columnas.}

\item{.wt}{Ponderador o expansor de los datos. Por defecto es NULL.}
}
\value{
tibble
}
\description{
Tablas de resultados ----------------------------------------------------
}
\details{
Generación de una data.frame con el número de casos y proporción de las distintas
variables de segmentos que se agregen en `...`.
}
