% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tablas.R
\name{tabla_segmentos}
\alias{tabla_segmentos}
\title{Tablas de resultados ----------------------------------------------------}
\usage{
tabla_segmentos(.data, ..., wt = NULL)
}
\description{
Tablas de resultados ----------------------------------------------------
}
