% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_vars_segmentos}
\alias{tabla_vars_segmentos}
\alias{tabla_vars_segmentos.data.frame}
\title{Tabla de porcentajes de variables según segmentos.}
\usage{
tabla_vars_segmentos(.df, .vars, .segmentos = NULL, ...)

\method{tabla_vars_segmentos}{data.frame}(
  .df,
  .vars,
  .segmentos = NULL,
  .wt = NULL,
  miss = NULL,
  total = FALSE,
  ...
)
}
\arguments{
\item{.df}{Un data.frame, tibble o tbl_svy}

\item{.vars}{tidyselect, lista de nombres de variables de las que se quiere saber su proporción de respuestas}

\item{.segmentos}{tidyselect, lista de nombres de variables de segmentación de las preguntas de `.vars`}

\item{...}{Parámetros adicionales que serán pasados a los métodos específicos}

\item{.wt}{name, nombre de la variable de ponderación}

\item{miss}{integers, Vector de valores que deben coniderarse como missings.}

\item{total}{logical, Si total = TRUE, se agrega el total para cada segmento.}
}
\value{
tibble
}
\description{
Obtiene porcentajes de respuestas de múltiples variables según múltiples segmentos.
Función genérica que funciona con data.frames regulares y objetos tbl_svy.
}
