% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_output.R
\name{tabla_vars_segmentos}
\alias{tabla_vars_segmentos}
\title{Tabla de porcentajes de variables según segmentos.

 Obtiene porcentajes de respuestas de múltiples variables según multiples segmentos.}
\usage{
tabla_vars_segmentos(.data, .vars, .segmentos, .wt = NULL,
  total = FALSE, miss = NULL)
}
\arguments{
\item{.data}{tibble}

\item{.vars}{vars(), lista de nombres de variables de las que se quiere saber su proporción de respuestas}

\item{.segmentos}{vars(), lista de nombres de variables de segmentación de las preguntas de `.vars`}

\item{.wt}{name, nombre de la variable de ponderación}

\item{total}{logical, Si total = TRUE, se agrega el total para cada segmento.}

\item{miss}{integers, Vector de valores que deben coniderarse como missings.}
}
\value{
tibble
}
\description{
Tabla de porcentajes de variables según segmentos.

 Obtiene porcentajes de respuestas de múltiples variables según multiples segmentos.
}
